requirements.txt
altgraph==0.17.3
anyio==3.6.2
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
comtypes==1.1.14
convertdate==2.4.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
exchange-calendars==4.2.6
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
jdcal==1.4.1
Jinja2==3.1.2
keyboard==0.13.5
korean-lunar-calendar==0.3.1
lunardate==0.2.0
lxml==4.9.2
MarkupSafe==2.1.2
mouse==0.7.1
MouseInfo==0.1.3
numpy==1.23.3
openpyxl==3.0.7
outcome==1.2.0
packaging==23.0
pandas==2.0.1
pandas-market-calendars==4.1.4
pefile==2023.2.7
Pillow==9.5.0
psutil==5.9.4
PyAutoGUI==0.9.53
pycparser==2.21
PyExcelerate==0.10.0
PyGetWindow==0.0.9
pyinstaller==5.10.0
pyinstaller-hooks-contrib==2023.2
pyinstaller-versionfile==2.1.1
pyluach==2.2.0
PyMeeus==0.5.12
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
python-telegram-bot==20.2
pytweening==1.0.4
pytz==2023.3
pywin32==306
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==6.0
requests==2.28.2
rfc3986==1.5.0
selenium==4.8.3
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
toolz==0.12.0
trio==0.22.0
trio-websocket==0.10.2
tzdata==2023.3
urllib3==1.26.15
workalendar==17.0.0
wsproto==1.2.0
xlrd==2.0.1
xlwings==0.30.4
xlwt==1.3.0
